requirements.txt
discord
python-dotenv
aiosqlite3
jishaku